requirements.txt
absl-py==0.13.0
alabaster==0.7.12
apex==0.1
appdirs==1.4.4
argon2-cffi==20.1.0
ascii-graph==1.5.1
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
audioread==2.1.9
Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.9.3
bleach==3.3.1
blis==0.7.4
boto3==1.18.5
botocore==1.21.5
brotlipy==0.7.0
cachetools==4.2.2
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.0
click==7.1.2
codecov==2.1.11
colorama==0.4.4
conda==4.10.3
conda-build==3.21.4
conda-package-handling==1.7.3
configparser==5.2.0
coverage==5.5
cryptography==3.4.7
cycler==0.10.0
cymem==2.0.5
Cython==0.28.4
DataProperty==0.52.0
debugpy==1.4.0
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.5
DLLogger==0.1.0
docker-pycreds==0.4.0
docutils==0.16
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
Flask==2.0.1
flatbuffers==2.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
glob2==0.7
google-auth==1.33.1
google-auth-oauthlib==0.4.4
graphsurgeon==0.4.5
grpcio==1.39.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.3.0
html2text==2020.1.16
httptools==0.2.0
hypothesis==4.50.8
idna==2.10
imageio==2.9.0
imagesize==1.2.0
inflect==5.3.0
iniconfig==1.1.1
ipdb==0.13.9
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
IQA-pytorch==0.1
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.2
jupyter-core==4.7.1
jupyter-tensorboard==0.2.0
jupyterlab==2.3.1
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupyterlab-widgets==1.1.1
jupyterthemes==0.20.0
jupytext==1.11.4
kiwisolver==1.3.1
lesscpy==0.15.0
libarchive-c==3.1
librosa==0.8.0
llvmlite==0.35.0
lmdb==1.2.1
Mako==1.1.4
Markdown==3.3.4
markdown-it-py==1.1.0
MarkupSafe==2.0.1
maskrcnn-benchmark==0.1
matplotlib==3.3.4
matplotlib-inline==0.1.2
mbstrdecoder==1.0.1
mccabe==0.6.1
mdit-py-plugins==0.2.8
mistune==0.8.4
mlperf-compliance==0.0.10
mock==4.0.3
msgfy==0.1.0
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.0
nltk==3.6.2
notebook==6.2.0
numba==0.52.0
numpy==1.19.5
nvidia-dali-cuda110==1.3.0
nvidia-dlprof-pytorch-nvtx==1.3.0
nvidia-dlprofviewer==1.3.0
oauthlib==3.1.1
onnx==1.10.2
onnxruntime==1.8.1
opencv-python==4.5.3.56
packaging==21.0
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pathtools==0.1.2
pathvalidate==2.4.1
pathy==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==21.1.3
pkginfo==1.7.1
pluggy==0.13.1
ply==3.11
polygraphy==0.31.1
pooch==1.4.0
preshed==3.0.5
prettytable==2.1.0
progressbar==2.5
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.7.0
pycocotools==2.0+nv0.5.1
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pycuda==2020.1
pydantic==1.8.2
pydot==1.4.2
pyflakes==2.1.1
Pygments==2.9.0
pynvml==8.0.4
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
pytablewriter==0.47.0
pytest==6.2.4
pytest-cov==2.12.1
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.18.0
python-hostlist==1.21
python-nvd3==0.15.0
python-slugify==5.0.2
pytools==2021.2.7
pytorch-quantization==2.1.0
pytorch-transformers==1.1.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.1.0
qtconsole==5.3.1
QtPy==2.1.0
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
revtok==0.0.3
rsa==4.7.2
ruamel-yaml-conda==0.15.80
s3transfer==0.5.0
sacrebleu==1.2.10
sacremoses==0.0.35
scikit-image==0.15.0
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.7.1
sentencepiece==0.1.96
sentry-sdk==1.8.0
setuptools==49.6.0.post20210108
shellingham==1.4.0
shortuuid==1.0.9
six==1.16.0
smart-open==5.1.0
smmap==5.0.0
snowballstemmer==2.1.0
SoundFile==0.10.3.post1
soupsieve==2.0.1
sox==1.4.1
spacy==3.1.1
spacy-legacy==3.0.8
Sphinx==4.1.1
sphinx-glpi-theme==0.3
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
srsly==2.4.1
subprocess32==3.5.4
subword-nmt==0.3.3
tabledata==1.2.0
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorrt==8.0.1.6
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
thinc==8.0.8
thop==0.0.31.post2005241907
threadpoolctl==2.2.0
toml==0.10.2
torch==1.10.1
torchtext==0.11.0a0
torchvision==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
typepy==1.2.0
typer==0.3.2
typing-extensions==3.10.0.0
uff==0.6.9
Unidecode==1.2.0
urllib3==1.26.6
uvicorn==0.14.0
uvloop==0.15.3
wandb==0.12.9
wasabi==0.8.2
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==9.1
Werkzeug==2.0.1
wheel==0.36.2
whitenoise==5.3.0
widgetsnbextension==3.6.1
wrapt==1.10.11
yacs==0.1.8
yaspin==2.1.0
